+++ web/default/src/components/data-table/core/table-skeleton.tsx
