+++ web/default/src/components/data-table/index.ts
@@ -16,27 +16,30 @@ along with this program. If not, see <https://www.gnu.org/licenses/>.
 
 For commercial licensing, please contact support@quantumnous.com
 */
-export { DataTablePagination } from './pagination'
-export { DataTableColumnHeader } from './column-header'
-export { DataTableViewOptions } from './view-options'
-export { DataTableToolbar } from './toolbar'
-export { DataTableBulkActions } from './bulk-actions'
+export { DataTablePagination } from './core/pagination'
+export { DataTableColumnHeader } from './core/column-header'
+export { DataTableViewOptions } from './toolbar/view-options'
+export { DataTableToolbar } from './toolbar/toolbar'
+export { DataTableBulkActions } from './toolbar/bulk-actions'
 export {
   StaticDataTable,
   type StaticDataTableColumn,
-} from './static-data-table'
-export { staticDataTableClassNames } from './static-data-table-classnames'
+} from './static/static-data-table'
+export { staticDataTableClassNames } from './static/static-data-table-classnames'
 export {
   DataTableRow,
   DataTableView,
   type DataTableColumnClassName,
   type DataTablePinnedColumn,
   type DataTableRenderRowHelpers,
-} from './data-table-view'
-export { MobileCardList } from './mobile-card-list'
-export { DataTablePage, type DataTablePageProps } from './data-table-page'
-export { useDataTable } from './use-data-table'
-export { useDebouncedColumnFilter } from './use-debounced-column-filter'
+} from './core/data-table-view'
+export { MobileCardList } from './layout/mobile-card-list'
+export {
+  DataTablePage,
+  type DataTablePageProps,
+} from './layout/data-table-page'
+export { useDataTable } from './hooks/use-data-table'
+export { useDebouncedColumnFilter } from './hooks/use-debounced-column-filter'
 
 export const DISABLED_ROW_DESKTOP =
   'bg-muted/85 hover:bg-muted [&>td:first-child]:border-l-muted-foreground/35 [&>td:first-child]:border-l-4 [&>td:first-child]:pl-1'
