+++ web/default/src/components/data-table/static/static-data-table.tsx
